Refactored project directory.
