+++ lib/pages/home/home_main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:portfolio_website/pages/home/home_page_one.dart';
-import 'package:portfolio_website/pages/home/home_page_three.dart';
-import 'package:portfolio_website/pages/home/home_page_two.dart';
+import 'package:portfolio_website/pages/home/home_page.dart';
+import 'package:portfolio_website/pages/projects/project_page_main.dart';
+import 'package:portfolio_website/pages/about_me/about_me_page.dart';
 import 'package:portfolio_website/providers/scroll_to_top_provider.dart';
 import 'package:provider/provider.dart';
